Gives Podmen not ass healing, but not terribly op either.
This makes it so it doesnt take 40 minutes to heal to full health, and at most takes about 20. by changing the health regen from 0.05 to 0.6 it greately improves it without making it OP in a fight. In most fights by the time their heal procs the fight would be over.